requirements.txt
Django
djangorestframework
python-decouple
PyJWT
django-cors-headers
